+++ tests/compilation_failures/core/extend_variable_and_then_set_it.rs
@@ -3,6 +3,6 @@ use preinterpret::*;
 fn main() {
     stream! {
         [!set! #variable = Hello]
-        [!set! #variable += World #(variable = [!stream! Hello2])]
+        [!set! #variable += World #(variable = %[Hello2])]
     }
 }