+++ tests/compilation_failures/expressions/embedded_variable_in_expression.rs
@@ -2,7 +2,7 @@ use preinterpret::*;
 
 fn main() {
     let _ = run! {
-        let partial_sum = [!stream! + 2];
+        let partial_sum = %[+ 2];
         5 #partial_sum
     };
 }