+++ tests/compilation_failures/expressions/variable_with_incomplete_expression.rs
@@ -2,6 +2,6 @@ use preinterpret::*;
 
 fn main() {
     let _ = stream!{
-        #(x = [!stream! + 1]; 1 x)
+        #(x = %[+ 1]; 1 x)
     };
 }